refactor: Domain Layer 세분화[entity / etc] (#5)
- Entity & 이외의 추가적인 Service를 위한 도메인 분리
