+++ src/main/java/com/sjiwon/studyholic/domain/entity/study/repository/StudyRepository.java
@@ -1,0 +1,8 @@
+package com.sjiwon.studyholic.domain.entity.study.repository;
+
+import com.sjiwon.studyholic.domain.entity.study.Study;
+import com.sjiwon.studyholic.domain.entity.study.repository.dsl.StudyQueryDslRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface StudyRepository extends JpaRepository<Study, Long>, StudyQueryDslRepository {
+}
