+++ src/main/java/com/sjiwon/studyholic/domain/entity/study/repository/dsl/StudyQueryDslRepository.java
@@ -1,0 +1,4 @@
+package com.sjiwon.studyholic.domain.entity.study.repository.dsl;
+
+public interface StudyQueryDslRepository {
+}
