+++ src/main/java/com/sjiwon/studyholic/domain/entity/studytag/repository/StudyTagRepository.java
@@ -1,0 +1,8 @@
+package com.sjiwon.studyholic.domain.entity.studytag.repository;
+
+import com.sjiwon.studyholic.domain.entity.studytag.StudyTag;
+import com.sjiwon.studyholic.domain.entity.studytag.repository.dsl.StudyTagQueryDslRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface StudyTagRepository extends JpaRepository<StudyTag, Long>, StudyTagQueryDslRepository {
+}
