+++ src/main/java/com/sjiwon/studyholic/domain/entity/studytag/repository/dsl/StudyTagQueryDslRepository.java
@@ -1,0 +1,4 @@
+package com.sjiwon.studyholic.domain.entity.studytag.repository.dsl;
+
+public interface StudyTagQueryDslRepository {
+}
