+++ src/main/java/com/sjiwon/studyholic/domain/entity/user/repository/dsl/UserQueryDslRepository.java
@@ -1,0 +1,4 @@
+package com.sjiwon.studyholic.domain.entity.user.repository.dsl;
+
+public interface UserQueryDslRepository {
+}
