+++ src/main/java/com/sjiwon/studyholic/domain/entity/userstudy/repository/UserStudyRepository.java
@@ -1,0 +1,8 @@
+package com.sjiwon.studyholic.domain.entity.userstudy.repository;
+
+import com.sjiwon.studyholic.domain.entity.userstudy.UserStudy;
+import com.sjiwon.studyholic.domain.entity.userstudy.repository.dsl.UserStudyQueryDslRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserStudyRepository extends JpaRepository<UserStudy, Long>, UserStudyQueryDslRepository {
+}
