+++ src/main/java/com/sjiwon/studyholic/domain/entity/userstudy/repository/dsl/UserStudyQueryDslRepository.java
@@ -1,0 +1,4 @@
+package com.sjiwon.studyholic.domain.entity.userstudy.repository.dsl;
+
+public interface UserStudyQueryDslRepository {
+}
