+++ src/main/java/com/sjiwon/studyholic/domain/etc/file/dto/UploadProfile.java
@@ -1,0 +1,11 @@
+package com.sjiwon.studyholic.domain.etc.file.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class UploadProfile {
+    private String uploadName;
+    private String storageName;
+}
