--- src/main/java/com/sjiwon/studyholic/domain/file/dto/UploadProfile.java
@@ -1,11 +1,0 @@
-package com.sjiwon.studyholic.domain.file.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class UploadProfile {
-    private String uploadName;
-    private String storageName;
-}
