--- src/main/java/com/sjiwon/studyholic/domain/study/repository/dsl/StudyQueryDslRepository.java
@@ -1,4 +1,0 @@
-package com.sjiwon.studyholic.domain.study.repository.dsl;
-
-public interface StudyQueryDslRepository {
-}
