--- src/main/java/com/sjiwon/studyholic/domain/studytag/repository/dsl/StudyTagQueryDslRepository.java
@@ -1,4 +1,0 @@
-package com.sjiwon.studyholic.domain.studytag.repository.dsl;
-
-public interface StudyTagQueryDslRepository {
-}
