--- src/main/java/com/sjiwon/studyholic/domain/user/repository/UserRepository.java
@@ -1,14 +1,0 @@
-package com.sjiwon.studyholic.domain.user.repository;
-
-import com.sjiwon.studyholic.domain.user.User;
-import com.sjiwon.studyholic.domain.user.repository.dsl.UserQueryDslRepository;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long>, UserQueryDslRepository {
-    Optional<User> findByLoginId(String loginId);
-    boolean existsByNickName(String nickName);
-    boolean existsByLoginId(String loginId);
-    boolean existsByEmail(String Email);
-}
