--- src/main/java/com/sjiwon/studyholic/domain/user/repository/dsl/UserQueryDslRepository.java
@@ -1,4 +1,0 @@
-package com.sjiwon.studyholic.domain.user.repository.dsl;
-
-public interface UserQueryDslRepository {
-}
