--- src/main/java/com/sjiwon/studyholic/domain/userstudy/repository/dsl/UserStudyQueryDslRepository.java
@@ -1,4 +1,0 @@
-package com.sjiwon.studyholic.domain.userstudy.repository.dsl;
-
-public interface UserStudyQueryDslRepository {
-}
